Page reloading when widgets moved
Seems localStorage shouldn't be changed from within $scope.updae();

Since other pages listen to the event and then themselves make a
localStorage change, the source page also gets the event. If 3 or more
pages are open, this could create an infinite loop.

One solution is to check to see if it needs to be changed rather than
always making the changes. Another solution is to add a new storage
version, update old values, and ensure new values are set properly.
Regardless, the simple check (without updating localStorage) is fine for
now.